Implemented local port randomisation for UDP and TCP sockets according to our latest discussion on the mailing list.
Each time a new socket is created the local port number in incremented by a 
more or less randomized value between 1 and 15 (the lowest 4 bit of 
NutGetMillis() | 1). The highest two bits are always set to 1.

This way a port range of 49152 to 65535 is used according to the IANA 
suggestions for ephemeral port usage.

From the security point of view this kind of local port randomisation still 
gives verry predictable results, but it is still better than just incrementing
and should help us to avoid problems as described in 

http://lists.egnite.de/pipermail/en-nut-discussion/2005-November/005290.html



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@4356 1ca430b3-8271-4a6b-9b0e-3dcac98a8924