The Cortex implementation used the System Core Debug unit CPU Cycle counter to implement a high precise NutMicroDelay() function.
Unfortunately on the LPC CPUs it seems that the Core Debug registers may not 
be user programmed which caused NutMicroDelay to freeze the system.

Curiously it works if a JTAG Debugger is connected.

As long as this function does not work reliable I did not enable it by default.

Use the new config option "NUT_MICRODELAY_CM3_CYCCNT" in the RTOS->TimerManagement
section to re-enable it.



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@4458 1ca430b3-8271-4a6b-9b0e-3dcac98a8924