Fixed problem with the tool path when building Nut/OS on Linux.
The PATH formatting is different between Windows and Linux so we also have
to differentiate the platform and set the PATH accordingly.

On Windows the PATH seperator is a ';', on Linux a ':' is used. 
Further more the environment variables are case sensitive on Linux, so we to 
use the all-upper-case PATH variable




git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@4594 1ca430b3-8271-4a6b-9b0e-3dcac98a8924