GPIO STM32/AVR: Allow to set the initial value of pin output data register
This helps to avoid glitches when initializing the port. E.g. a pin pulled-up
outside when coming out of reset mostly has the output register reset. When
this line is configured multidrive, the line will be driven low as long as the
output register is not set. This may result in unwanted behaviour.

Needs to be considered for the other archtectures too

git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@4738 1ca430b3-8271-4a6b-9b0e-3dcac98a8924