Implemented posix style file descriptor mapping, where the file descriptors are integers used as array index. Until now we mapped a pointer directly to an integer and used it as filedescriptor. This way we could crash the system when passing invalid file descriptors
Further more the typical posix select implementation needs the filedescriptors
to be indexes into an array with a given maximum number of filedescriptors.

Currently I limited the maximum number of concurrently opened files to 16 by 
default. I reused the FOPEN_MAX value for defining the limit. FOPEN_MAX was
used as maximum number of concurrently opened _streams_ and was set to 8
until now.

The new implementation needs FOPEN_MAX bytes of statically allocated memory
for holding the list of filedescriptors.



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/branches/devnut_select_or@5002 1ca430b3-8271-4a6b-9b0e-3dcac98a8924