Exchanged the so_device <-> so_next entries in the TCP socket struct and also vdv_zero <-> vdv_next in the NUTVIRTUALDEVICE struct.
Why?

In one of the recent checkins, I removed the nf_next entry from NUTFILE struct
as it was used nowhere in the code.

But some of the nut/crt/* functions try to check, if they are called with a 
NUTFILE based filedescritor, or of they are called with a file descritor
pointing to a TCPSOCK struct. This check is done by testing for 
vdv_zero == NULL, while the NUTFILE or TCPSOCK pointer is casted to
NUTVIRTUALDEVICE. In other words: we removed one entry from NUTFILE (nf_next),
so we have to change to order of the variables in the casted structs 
accordingly.



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/branches/devnut_select_or@5065 1ca430b3-8271-4a6b-9b0e-3dcac98a8924