Applied Adeels patch. Work arround for an errata:
41.4.8.7

SPI: Software Reset must be Written Twice

If a software reset (SWRST in the SPI Control Register) is performed,
the SPI may not work properly (the clock is enabled before the chip select.)

Problem Fix/Workaround:

The SPI Control Register field, SWRST (Software Reset) needs to be
written twice to be correctly set.


Thanks Adeel for debugging this!




git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@5094 1ca430b3-8271-4a6b-9b0e-3dcac98a8924