Implemented a new make target "check-licenses", which can be used to add checks, if all configuration options are properly set to completely compile the Nut/OS code.
Currently it us used by the STM32 platform, to check wether the propriatory
STM licences are accepted in the configurator. If not, a warning is printed
out during compilation to notify the user about possible build errors 
caused by missing functions, that are only available if these licenses are 
accepted.

To implement this feature, a few changes / additions to the configurator 
scripts and the Makefiles were needed:

Configurator:

The Configurator options that shall be checked need to write an evironment 
variable into Nutconf.mk, same is done when selecting STM32 as architecture.

Makefiles:

A new default target was created in Makevars.all (as it needs to stay on top 
of the evaluated make targets) which is called "default". 
Target "default" depends on "check-licenses all" which causes the license check
to be run onece before the compilation process starts.

The check-license target now prints out a message, which is created as 
a variable in Makerules.licenses.




git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@5257 1ca430b3-8271-4a6b-9b0e-3dcac98a8924