!! This is a major change in the way, the system time is calculated !!
The old way of calculating the current time was based on the system tick 
counter. Unfortunately this introduced bugs when this counter value overflowed. 

On most systems (with 1Khz system tick timer) this happened after 49,7 days. 
If the system used an RTC, this bug did not show up in most cases, as the 
time was then queried from the RTC.

Another problem was, that the time could not be corrected with an accuracy 
of more than 1 second, as the offset to epoc was calculated as a value of 
second resolution.

The new time API re-invented the way the time is handled and calculated.
 
Internally the time is now stored as a timeval struct, which allowes a 
a resolution up to 1 µs. 
The time is now always updated by the system tick timer interrupt and is 
always hold as software clock. It is _not_ automatically synced with 
the RTC any more. Syncing with a RTC to be done manually by the user!
 
This way we gain a much higher accuracy, as long, as the software clock 
is accurate. In other words, the time depends on the system crystal accuracy 
and time will not be updated, if the CPU is put to deep sleep mode. 

Further more we assume a system tick timer running at a multiple of a full µs.
Uneven system tick timer frequencies will result in an increasing error of 
the software clock.

For any application that need to rely on the old API a new configurator option
has been introduced: "NUT_USE_OLD_TIME_API"

Enable this option, if you would like to stay with the old API. This might be 
usefull if you design a low-power system, where the CPU is sleeping most of 
the time and shall not be woken up regularly, or if you use an odd crystal 
frequency which can not provide a system tick timer running at an even 
multiple of 1µs.



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@5261 1ca430b3-8271-4a6b-9b0e-3dcac98a8924