Last checking breaks GPIO on NXP LPC17xx Reverted last checkin
My initial version implemented GpioPinSet by writing to the FIOSET register 
for high values and FIOCLR for low values. The last checking replaced it by 
writing either 1 or 0 to the FIOSET register, which will not work as you 
can only write ones to these registers, writing 0 won't modify the register.

According to the datasheet it would be possible to directly write the 
final value to the FIOPIN register. Due to a silicon bug this is highly 
discouraged to do so, using the bit banding register access.

If writing two values using bit band access to the FIOPIN register directly 
one after the other, then the first access might get missed.

Conclusion:

We have to stay with writing to FIOSET and FIOCLR



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@5423 1ca430b3-8271-4a6b-9b0e-3dcac98a8924