Fixed race condition in TCP statemachine and modified the way sockets are destroyed, added garbage collector.
Problem: Race condition in tcpSocketList handling

NutTcpDestroySocket() is called from different places within the TCP 
statemachine, as well as indirectly from NutTcpCloseSocket() in some
situation.

NutTcpDestroySocket() first removed the given socket from
tcpSocketList and then frees the allocated memory.

Problem:
========

tcpSocketList is a linked list of the currently existing TCP sockets. A
call to NutTcpDestroySocket() removes the given socket from this list by
re-linking the so_next-> pointer. The socket which is removed can stay
at any place in the list. So it is not necessary the first or the last
entry.

The problem is, that tcpSocketList is used at several places in Nut/OS
without any kind of locking. Mainly it is used in THREAD(NutTcpSm, arg)
to iterate over its entries.

Race condition:
===============

While iterating over the socket list, the socket list itself can be modified,
which means that in some situations, it could happen that the
socket pointer is freed and unlinked by a call to NutTcpDestroySocket(),
while it is used later in the loop again.

Solution:
=========

* Instead of calling NutTcpDestroySocket() from several places within
  the statemachine (or from NutTcpCloseSocket()), a new socket state is
  introduced: TCPS_DESTROY. If the socket shall be destroyed, the socket
  state will be changed to TCPS_DESTROY and the flag tcp_run_gc is set.

* The statemachine was modified accordingly to handle this new
  state correctly and reject any communication in this state.

* A new function NutTcpGarbadgeCollect() is added, which is called by the 
  TCP statemachine, while it is idle, provided tcp_run_gc is set. 
  This function itereates over all sockets and unlinks / frees ressources 
  of all sockets in TCPS_DESTROY state.

* The garbage collector runs outside of any loop iterating over the socket
  list and does not have any scheduling point.



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@5541 1ca430b3-8271-4a6b-9b0e-3dcac98a8924