Added generic cryptographic and TLS client / server library ===========================================================
The code is based on the axTLS library (http://axtls.sourceforge.net/). 
The axTLS embedded SSL project is a highly configurable client/server 
TLSv1 SSL library designed for platforms with small memory requirements.

The code is based on axTLS 1.4.9

It has been ported to Nut/OS. It was tested only on 32 bit platforms,
but should run well ob AVR as well, while on AVR the available RAM
might be the limiting factor.

The ressource requirements are moderate. The library itsel is about 40-70K
in size. During runtime a minimum of 64K of RAM are needed. The SSL block size
is up to 16K, so during runtime each SSL session takes more than 16K +x.

The stack size should be more than 2K for the SSL session thread.

The session establishment is astonishing fast as the needed private key
must be supplied during compile time or at run-time (e.g. on a sd-card)

Lots of omtimizations for Nut/OS have just been implemented, while there is
still lots of space for further optimizations. Especially to reduce the
needed stack size.

The crypto library provides the following algorithms:
=====================================================

- generic large integer algebra (needed e.g. for rsa)
- aes 128 / 256
- md2
- md5
- sha1
- rc4
- rsa
- hmac-md5, hmac-sha1

The TLS library provides the following features:
================================================

- TLS library with client and server support
- TLSv1.0/1.1, fallback to ssl 2.3 handshake
- Session resumption for high SSL performance.
- very simple API
- Peer client/server verification, certificate chaining, 
  session renegotiation.
- X509v1, PKCS#8, PKCS#12 keys/certificates in DER/PEM format
- Runtime certificate generation

Not supported:

- SSLv2 or SSLv3 (but supports v23 client hello).
- Anonymous or Ephemeral DH.
- DSA or ECC.
- DES/3DES ciphers.
- Key generation (however certificate generation is supported).




git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@5597 1ca430b3-8271-4a6b-9b0e-3dcac98a8924