Added experimental stack backtrace code, which can be enabled for easier debugging exceptions.
Enable DEBUG_BACKTRACE in the configurator to enable the backtrace 
feature.

In case of an exception the last 128 words on the stack will
be printed out and a call trace will be generated and printed out
(Address of the function entry and offset in the function).
Together with the .map file and / or the assembler listing
the call trace allows you to trace back your code flow in
which led to the exception. The assembler listing can be
created by 'arm-none-eabi-objdump -d myfile.elf'

This function is highly experimental and relies on a very
defined function entry code and existance of a stack
frame pointer
To ensure the correct function layout this option disables
compiler optimisation and enables a stack frame pointer
generation. The following compiler options will be set:

-O0 -mtpcs-frame -mtpcs-leaf-frame -fno-omit-frame-pointer

This will enlarge the resulkting binary size significantly
any may influence execution performance.

The backtrace will only work for functions compiled with
above compiler parameters. E.g. it will very likely not 
workfor any newlib functions.

The backtrace needs a frame pointer to be saved on each function
entry. With the above compiler settings, r7 will be used as 
frame poiter. So the default exception handler code has been
altered to pass r7 as third parameter to CortexRegDump().
CortexRegDump() itself has been extended to print out the 
call trace...

As mentioned above, this code is quite experiemental but may
help to debug your applications, if the application is just 
somewhere in the field.



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@5684 1ca430b3-8271-4a6b-9b0e-3dcac98a8924