lpc177x_8x_mci.c: Fixed a dodgy bug in the SD-Card driver
When reading or writing a block from/to SD-Card, the mci FIFO is used. It is
split into two half. As soon as the FIFO is half-empty, an interrupt is 
generated to re-start filling.

Even if the block(s) are yet completely written, the interrupt is generated 
two more times (one for half empty, one for full empty). 

As it was not checked, if the buffer was yet completely read or written,
it was either read or written behind the end of the buffer.

Fixed by checking the read / written number of bytes and disabling the 
interrupt until next read / write command. 



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@5760 1ca430b3-8271-4a6b-9b0e-3dcac98a8924