Implemented configurable system wide TCP statemachine incomming buffer limit.
The TCP statemachine uses HEAP buffer to buffer incomming TCP packets. This
is independend on socket connections. These packets are buffered, until they
are evaluated and passed to the application. This can may cause massive memory
drops. A limit of 2048 Bytes had beed implemented hardcoded, this have been
made configurable.

For applications with enough memory, a limit of 64K or higher seems 
reasonable.



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6140 1ca430b3-8271-4a6b-9b0e-3dcac98a8924