Limit the socket receive buffer to a configurable value.
A new configurator option (Network / TCP / TCP_SOCK_RXBUF_LIMIT) has been
introduced, which can be used to limit the socket receive buffer. 
The value is a multiple of the software configured socket receive buffer 
(SO_RCVBUF).       

Under normal conditons no buffering beyond the SO_RCVBUF should occur, 
but some TCP implementations seem to not follow the TCP RFCs correctly
and continue sending bytes, even when sending a zero-window ack. So this
option is for limitting the used HEAP in these situations.

If configured to a value > 0, we allow to buffer a maxmimum of
TCP_SOCK_RXBUF_LIMIT * SO_RCVBUF bytes, until we drop further incomming 
packets for this socket. 
A value of '0' (default) will result in the known default behaviour, 
with an unlimited socket receive buffer.



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6141 1ca430b3-8271-4a6b-9b0e-3dcac98a8924