Reverted commit r6110
In r6146 the AVR compiler issue (for crypto/sha512) had been worked around.
Now sha512.c will be compiled by default again.

Disabling sha512 compilation is not a good idea anyways, as it is needed
for tls/x509.c for example. If sha512.c is not compiled, linking of 
applications that need the TLS stack will fail.



git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6147 1ca430b3-8271-4a6b-9b0e-3dcac98a8924