Make TCP advance package buffer limit configurable (TCP_ADV_MAX)
The tcp advance package buffer had been globally limited to 3216 bytes, which
is the default window size (6 x 536 bytes).

This is global limit, for all sockets together. It turned out, that the limit
was set too low for lots of applications, resulting in unneded drops of 
incomming packages. Which finally slowed down stream recovery on unreliable
tcp connections.

This setting have been made configurable. The default setting is still 3216 
bytes.

Setting this value to 0 disables the buffer limit.

The higher this value (or if it is disabled) the faster the tcp stream recovery
will work on unreliable connections.


git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6286 1ca430b3-8271-4a6b-9b0e-3dcac98a8924