Fix bitband access issues to the GPIO registers by adding a NOP to some macros
We need to add _NOP() instructions to some of the functions as the GPIO_SET_x
and GPIO_INPUT and GPIO_OUTPUT macros set the related bits using the bitband
area. Unfortunately the bitbanding access to the registers seems somewhat
broken in the LPC17xx and the operations are not that atomic as one would
expect. If do two consecutive bitband writes to the GPIO pin registers, the
first one could get lost, if there is no NOP() between the two writes.

git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6290 1ca430b3-8271-4a6b-9b0e-3dcac98a8924