qnutconf: Rework defaults.
If a .conf file is given, look if the conf file resides in a Nut/OS tree.
If yes, use that tree as default for src_dir. In that case, use
nut/../<basename(xxx.conf)> as default for bld_dir  and app_dir.
If conf file is outside source tree, use <pathname(xxx.conf)>/nut(app|bld)
as defaults for bld_dir and app_dir.

git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6549 1ca430b3-8271-4a6b-9b0e-3dcac98a8924