stm32: Allow to run from ram, with restart possibility.
Storage area for initialized variables is doubled. On inital start,
initialization values are copied from the working area to a safe
storage area and on next warm starts the working area is initialised
from the safe storage.
Magic variable for this purpose are stored in space over the last vector in
RAM.
To allow some program in flash to jump to the program loaded to RAM,
initial stack is placed direct after the area for vectors in RAM, or at
start of RAM is vectors are places elsewhere. Otherwise program in flash
starting program in RAM might overwrite variables.

git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6638 1ca430b3-8271-4a6b-9b0e-3dcac98a8924