Fix memory leak in timer.c, introduced in revision 6796
The new API to to (re)use pre-defined timer structs introduced a new flag
variable. Depending on this flag, the timer struct is freed or not.

But the flag was not set to 0x00 when allocating a new timer struct. Depeding
on the (random) memory content of the freshly allocated memory the flag was
some random value, so in most situations a timer buffer was never freed again,
causing a serious memory leak over the time.

git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6927 1ca430b3-8271-4a6b-9b0e-3dcac98a8924